+++ src/app/hechizos/Hechizo.java
@@ -5,11 +5,11 @@ package app.hechizos;
  */
 public class Hechizo {
     public String nombre;
-    public int nivelDaño;
-	public int nivelCuracion;
+    public double nivelDaño;
+	public double nivelCuracion;
 	public int nivelEnergiaMagica;
 
-    public Hechizo(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica) {
+    public Hechizo(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica) {
         this.nombre = nombre;
         this.nivelDaño = nivelDaño;
 		this.nivelCuracion = nivelCuracion;
@@ -28,7 +28,7 @@ public class Hechizo {
 		return nivelDaño;
 	}
 
-	public void setNivelDaño(int nivelDaño) {
+	public void setNivelDaño(double nivelDaño) {
 		this.nivelDaño = nivelDaño;
 	}
 
@@ -36,7 +36,7 @@ public class Hechizo {
 		return nivelCuracion;
 	}
 
-	public void setNivelCuracion(int nivelCuracion) {
+	public void setNivelCuracion(double nivelCuracion) {
 		this.nivelCuracion = nivelCuracion;
 	}
