+++ src/app/hechizos/ataque/PetrificusTotalus.java
@@ -4,7 +4,7 @@ package app.hechizos.ataque;
  * PetrificusTotalus
  */
 public class PetrificusTotalus extends HechizoAtaque {
-    public PetrificusTotalus(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public PetrificusTotalus(String nombre,  double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
     public PetrificusTotalus() {
