+++ src/app/hechizos/curacion/VulneraSanentur.java
@@ -5,7 +5,7 @@ package app.hechizos.curacion;
  */
 public class VulneraSanentur extends HechizoCuracion {
 
-    public VulneraSanentur(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public VulneraSanentur(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
     public VulneraSanentur() {
