+++ src/app/hechizos/defensa/CaveInmicum.java
@@ -5,7 +5,7 @@ package app.hechizos.defensa;
  */
 public class CaveInmicum extends HechizoDefensa {
 
-    public CaveInmicum(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public CaveInmicum(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
     public CaveInmicum() {
