+++ src/app/hechizos/defensa/Protego.java
@@ -5,7 +5,7 @@ package app.hechizos.defensa;
  */
 public class Protego extends HechizoDefensa {
 
-    public Protego(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public Protego(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
     public Protego() {
