+++ src/app/hechizos/imperdonable/AvadaKedavra.java
@@ -5,7 +5,7 @@ package app.hechizos.imperdonable;
  */
 public class AvadaKedavra extends HechizoImperdonable {
 
-    public AvadaKedavra(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica) {
+    public AvadaKedavra(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
