+++ src/app/hechizos/imperdonable/HechizoImperdonable.java
@@ -7,7 +7,7 @@ import app.hechizos.Hechizo;
  */
 public class HechizoImperdonable extends Hechizo{
 
-    public HechizoImperdonable(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public HechizoImperdonable(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
 }
