+++ src/app/hechizos/ocio/HechizoOcio.java
@@ -9,7 +9,7 @@ import app.hechizos.Hechizo;
  */
 public class HechizoOcio extends Hechizo {
 
-    public HechizoOcio(String nombre, int nivelDaño, int nivelCuracion, int nivelEnergiaMagica ) {
+    public HechizoOcio(String nombre, double nivelDaño, double nivelCuracion, int nivelEnergiaMagica ) {
         super(nombre, nivelDaño, nivelCuracion, nivelEnergiaMagica);
     }
